Diverged branches for paper figures and velocity input noise analysis

+++ grid_cell_model/simulations/007_noise/simulation_velocity.py
@@ -1,7 +1,7 @@
 #
-#   simulation_stationary.py
+#   simulation_velocity.py
 #
-#   Main simulation run: Simulation of a stationary bump.
+#   Main simulation run: Bump velocity estimation.
 #
 #       Copyright (C) 2012  Lukas Solanka <l.solanka@sms.ed.ac.uk>
 #       
@@ -18,53 +18,51 @@
 #       You should have received a copy of the GNU General Public License
 #       along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
+import numpy as np
+from os.path import exists
 from numpy.random       import choice
 
 from models.parameters  import getOptParser
-from models.gc_net_nest import BasicGridCellNetwork
+from models.gc_net_nest import ConstantVelocityNetwork
 from data_storage       import DataStorage
 
 
 parser          = getOptParser()
+parser.add_option("--IvelMax",      type="float", help="Max constant velocity current input (pA)")
+parser.add_option("--dIvel",        type="float", help="Constant velocity current input step (pA)")
+
 (options, args) = parser.parse_args()
 
+
+output_fname = "{0}/{1}job{2:05}_output.h5".format(options.output_dir,
+        options.fileNamePrefix, options.job_num)
+if (exists(output_fname)):
+    print('Output file exists, doing nothing')
+    exit(0)
+
+# Run the simulation
+d = DataStorage.open(output_fname, 'w')
 
 out = []
 overalT = 0.
 ################################################################################
 for trial_idx in range(options.ntrials):
     print("\n\t\tStarting trial no. {0}\n".format(trial_idx))
-    ei_net = BasicGridCellNetwork(options, simulationOpts=None)
-    
-    const_v = [00.0, 0.0]
-    ei_net.setConstantVelocityCurrent_e(const_v)
-    
-    
-    stateRecF_e = choice(ei_net.E_pop, options.gammaNSample, replace=False)
-    stateRecF_i = choice(ei_net.I_pop, options.gammaNSample, replace=False)
-    
-    stateMonF_e_params = {
-            'withtime' : False,
-            'interval' : options.sim_dt*10,
-            'record_from' : ['I_clamp_GABA_A']
-    }
-    stateMonF_i_params = dict(stateMonF_e_params)
-    stateMonF_i_params['record_from'] = ['I_clamp_AMPA', 'I_clamp_NMDA']
+    trialOut = {}
+    IvelVec = np.arange(0.0, options.IvelMax + options.dIvel, options.dIvel)
+    trialOut['IvelVec'] = IvelVec
+    trialOut['IvelData'] = []
+    for Ivel in IvelVec:
+        const_v = [Ivel, 0.0]
+        ei_net = ConstantVelocityNetwork(options, simulationOpts=None, vel=const_v)
 
-    stateMonF_e = ei_net.getGenericStateMonitor(stateRecF_e,
-            stateMonF_e_params, 'stateMonF_e')
-    stateMonF_i = ei_net.getGenericStateMonitor(stateRecF_i,
-            stateMonF_i_params, 'stateMonF_i')
-    
-    ei_net.simulate(options.time, printTime=options.printTime)
-    ei_net.endSimulation()
-    out.append(ei_net.getAllData())
-    constrT, simT, totalT = ei_net.printTimes()
-    overalT += totalT
+        ei_net.simulate(options.time, printTime=options.printTime)
+        ei_net.endSimulation()
+        trialOut['IvelData'].append(ei_net.getMinimalSaveData())
+        constrT, simT, totalT = ei_net.printTimes()
+        overalT += totalT
+    out.append(trialOut)
 
-output_fname = "{0}/{1}job{2:05}_output.h5".format(options.output_dir,
-        options.fileNamePrefix, options.job_num)
-d = DataStorage.open(output_fname, 'w')
 d["trials"] = out
 d.close()
 print "Script total run time: {0} s".format(overalT)
